+++ shoppingmall-back/src/main/java/com/shoppingmallcoco/project/entity/order/OrderItem.java
@@ -16,20 +16,20 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "orderitem_seq_gen")
     @SequenceGenerator(name = "orderitem_seq_gen", sequenceName = "orderitem_SEQ", allocationSize = 1)
     @Column(name = "orderitemNo")
-    private Integer orderItemId;
+    private Long orderItemId;
 
     @Column(name = "prdNo", nullable = false)
-    private Integer prdNo;
+    private Long prdNo;
 
     // --- [필수 추가] 옵션 번호 ---
     @Column(name = "OptionNo", nullable = false) // (DB 컬럼명: OptionNo)
-    private Integer optionNo;
+    private Long optionNo;
 
     @Column(name = "price", nullable = false) // (DB 컬럼명 수정: price)
-    private Integer orderPrice;
+    private Long orderPrice;
 
     @Column(name = "qty", nullable = false) // (DB 컬럼명 수정: qty)
-    private Integer orderQty;
+    private Long orderQty;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "orderNo")
